Clean up a bit:
No need to have the shebang line in module files.

Shebang lines (#!/usr/bin/perl, #!/bin/bash, etc.) are used by the shell
when it opens the script to know what to run on it.

If you call a script with "./myscript", and it's not a compiled program,
the shell won't know how to run it. It will open the file and search the
first 10 lines or so a line that says "#!...", which means "use this in
order to run this file".

Example:

    $ cat > myscript.pl
    #!/usr/bin/perl
    print "hello, world\n";
    ^D (exit)

    # if we don't do this, we can't do "./myscript.pl"
    chmod +x myscript.pl

    # now when we do ./myscript.pl it will look for #!...
    # and it will find "/usr/bin/perl
    # so, the following line...
    ./myscript.pl

    # will be equivalent to:
    /usr/bin/perl myscript.pl

That's exactly how that works. Since modules are loaded by perl already,
there's no need to give hints for the shell. :)